requirements.txt
# Python 3.13 required
Pillow>=10.0.0
numpy>=1.26.0
instagrapi>=2.0.0
schedule>=1.2.0
